pyproject.toml
# Poetry
[tool.poetry]
name = "fl-components"
description = "Project containing the components fl-client and fl-server"
authors = [
    "Borja Arroyo Galende <borja.arroyog@upm.es>"
]
readme = "README.md"
package-mode = false
version = "0.0.0"

[tool.poetry.dependencies]
python = "^3.11"
flwr = "^1.9"
pandas = "^2.2"
mlflow-skinny = "^2.11"
numpy = "^1"

[tool.poetry.group.fl-client.dependencies]
torch = "^2.2"

[tool.poetry.group.fl-server.dependencies]

[tool.poetry.dev-dependencies]
pytest = "^8.2"
pytest-env = "^1.1"
pytest-cov = "^5.0"
ruff = "^0.5"
mypy = "^1.10"

[tool.mypy]
python_version = "3.11"

[[tool.mypy.overrides]]
module = ["mlflow.*", "cloudpickle.*"]
ignore_missing_imports = true


[tool.poetry-dynamic-versioning]
enable = true

# Test
[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[tool.pytest_env]
MLFLOW_URL = { value = "http://localhost:30246", skip_if_set = true }
NODE_NAME = { value = "test", skip_if_set = true }
DATA_PATH = { value = "common/data/iris.csv", skip_if_set = true }

[tool.pytest_coverage.run]
omit = ["tests/*"]

[tool.ruff]
target-version = "py311"

# Build system
[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
